requirements.txt
Pillow
numpy
sentence-transformers
torch
bitarray
pyflann
scikit-learn
tqdm